requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astro-sdk-python
dbt-core
apache-airflow[google]
astronomer-cosmos
